textapp/requirements.txt
Django==5.2.4
python-dotenv
transformers
torch
google-generativeai
SpeechRecognition
pydub
gtts
django-allauth